requirements.txt
boto3==1.4.4
botocore==1.10.12
PyYAML==3.12
click==6.7
virtualenv==15.1.0
colorama==0.3.7
chalice==1.6.0
